chaincfg: Move testnet3 code to testnet files.
This moves the various code that is specific to testnet to the
testnetparams.go file to keep the information housed together.  Along
the same lines, it moves the tests specific to the testnet params to
testnetparams_test.go.

It also unexports the block one ledger definition since it should only
be exposed via the params.
