+++ chaincfg/genesis.go
@@ -12,28 +12,6 @@ import (
 	"github.com/decred/dcrd/wire"
 )
 
-// TestNet3 ------------------------------------------------------------------------
-
-// testNet3GenesisBlock defines the genesis block of the block chain which
-// serves as the public transaction ledger for the test network (version 3).
-var testNet3GenesisBlock = wire.MsgBlock{
-	Header: wire.BlockHeader{
-		Version:      6,
-		PrevBlock:    chainhash.Hash{},
-		MerkleRoot:   genesisCoinbaseTx.TxHash(),
-		Timestamp:    time.Unix(1533513600, 0), // 2018-08-06 00:00:00 +0000 UTC
-		Bits:         0x1e00ffff,               // Difficulty 1 [000000ffff000000000000000000000000000000000000000000000000000000]
-		SBits:        20000000,
-		Nonce:        0x18aea41a,
-		StakeVersion: 6,
-	},
-	Transactions: []*wire.MsgTx{&genesisCoinbaseTx},
-}
-
-// testNet3GenesisHash is the hash of the first block in the block chain for the
-// test network (version 3).
-var testNet3GenesisHash = testNet3GenesisBlock.BlockHash()
-
 // SimNet -------------------------------------------------------------------------
 
 var simNetGenesisCoinbaseTx = wire.MsgTx{
